Reorganization
Reorganized some of the code to consolidate where there will be repeated logic.
